move to server-common
